[WAGON-221] removeCheckoutDirectory throws NPE. Applied Null check to prevent invalid NPE.
git-svn-id: https://svn.apache.org/repos/asf/maven/wagon/trunk@739606 13f79535-47bb-0310-9956-ffa450edef68

+++ wagon-providers/wagon-scm/src/main/java/org/apache/maven/wagon/providers/scm/ScmWagon.java
@@ -175,6 +175,11 @@ public class ScmWagon
     private void removeCheckoutDirectory()
         throws ConnectionException
     {
+        if( checkoutDirectory == null )
+        {
+            return; // Silently return.
+        }
+
         try
         {
             FileUtils.deleteDirectory( checkoutDirectory );
